Modified the type of variable in the function SHgetUnrLen from int to long long. Because when we try to compile large size of array,e.g. 65535*65535, in the currey status the maximum of number int can represent is 2^32-1, which is smaller than 65535*65535. This will cause many probelm in the later process.

+++ src/libsac2c/constants/shape.h
@@ -32,7 +32,7 @@ extern void SHtouchShape (shape *shp, info *arg_info);
 
 extern int SHgetDim (shape *shp);
 extern int SHgetExtent (shape *shp, int dim);
-extern int SHgetUnrLen (shape *shp);
+extern long long SHgetUnrLen (shape *shp);
 
 extern int SHsubarrayDim (shape *shp, int n);
